Configuration Bits for TMR0 & Interrupt

+++ motor.c
@@ -4,69 +4,54 @@
 static uint8_t stepState, masterCount;
 static Direction directionState;
 
-static void initStep(void)
-{
+static void initStep(void) {
     resetStep();
     resetMasterCount;
 }
 
-uint8_t currentStep(void)
-{
+uint8_t currentStep(void) {
     return stepState;
 }
 
-void nextStep(Direction state)
-{
-    if (state == FORWARD)
-    {
-        if (stepState == 3)
-        {
+void nextStep(Direction state) {
+    if (state == FORWARD) {
+        if (stepState == 3) {
             stepState = 0;
         }
-        else
-        {
+        else {
             stepState++;
         }
     }
-    else
-    {
-        if (stepState == 0)
-        {
+    else {
+        if (stepState == 0) {
             stepState = 3;
         }
-        else
-        {
+        else {
             stepState--;
         }
     }
 }
 
-void resetStep(void)
-{
+void resetStep(void) {
     stepState = 0;
 }
-uint8_t currentMasterCount(void)
-{
+int currentMasterCount(void) {
     return masterCount;
 }
 
-static void incrementMasterCount(void)
-{
+static void incrementMasterCount(void) {
     masterCount++;
 }
 
-static void decrementMasterCount(void)
-{
+static void decrementMasterCount(void) {
     masterCount--;
 }
 
-static void resetMasterCount(void)
-{
+static void resetMasterCount(void) {
     masterCount = 0;
 }
 
-void initMotor(void)
-{
+void initMotor(void) {
     //Sent all motor ports to OUTPUT mode
     MOTOR_TRIS_0 = 0;
     MOTOR_TRIS_1 = 0;
@@ -86,46 +71,38 @@ void initMotor(void)
 }
 
 /* Write function for the directional state */
-void setMotorDirection(Direction state)
-{
+void setMotorDirection(Direction state) {
     directionState = state;
 }
 
 /* Read function for the directional state */
-Direction currentMotorDirection(void)
-{
+Direction currentMotorDirection(void) {
     return directionState;
 }
 
-void startMotor(void)
-{
+void startMotor(void) {
 
 }
 
-void stopMotor(void)
-{
+void stopMotor(void) {
 
 }
 
 /* This function is for the interrupt to use to control the motor */
-void stepMoter(void)
-{
-    if (currentMotorDirection == FORWARD)
-    {
+void stepMoter(void) {
+    if (currentMotorDirection == FORWARD) {
         nextStep(FORWARD);
         updateMotor(currentStep());
         incrementMasterCount();
     }
-    else
-    {
+    else {
         nextStep(BACKWARD);
         updateMotor(currentStep());
         decrementMasterCount();
     }
 }
 
-void holdMotor(void)
-{
+void holdMotor(void) {
     //Send a HIGH to all four pins of the stepper motor
     MOTOR_0 = 1;
     MOTOR_1 = 1;
@@ -133,8 +110,7 @@ void holdMotor(void)
     MOTOR_3 = 1;
 }
 
-void releaseMotor(void)
-{
+void releaseMotor(void) {
     //Send a LOW to all four pins of the stepper motor
     MOTOR_0 = 0;
     MOTOR_1 = 0;
@@ -142,10 +118,8 @@ void releaseMotor(void)
     MOTOR_3 = 0;
 }
 
-void updateMotor(uint8_t step)
-{
-    switch(step)
-    {
+void updateMotor(uint8_t step) {
+    switch(step) {
         case 0:
             releaseMotor();
             MOTOR_0 = 1;
